+++ lib/cldr/territory.ex
@@ -15,6 +15,7 @@ defmodule Cldr.Territory do
     {k, v} when is_list(v) -> {:"#{k}", Enum.map(v, &:"#{&1}")}
     {k, v} -> {:"#{k}", :"#{v}"}
   end)
+  @territory_codes Cldr.Config.territory_codes()
 
   @doc """
   Returns a list of available styles.
@@ -740,6 +741,16 @@ defmodule Cldr.Territory do
 
       {:ok, result}              -> result
     end
+  end
+
+  @doc """
+  Returns a mapping of territory codes (ISO 3166 Alpha-2 codes) to
+  their respective ALpha-3, FIPS 10 and numerica codes (where available).
+
+  """
+  @doc since: "2.12.0"
+  def territory_codes do
+    @territory_codes
   end
 
   @doc """